feat: BulkEditNodesJob allow value depending remove for Remove mode

+++ Backend/src/org/edu_sharing/repository/server/jobs/quartz/BulkEditNodesJob.java
@@ -97,7 +97,7 @@ public class BulkEditNodesJob extends AbstractJob{
 		ReplaceMapping,
 		@JobFieldDescription(description = "Currently Unsupported")
 		Append,
-		@JobFieldDescription(description = "Remove the property")
+		@JobFieldDescription(description = "Remove the property. Use with searchtoken: one value must be equal, than the property is removed.")
 		Remove
 	};
 
@@ -137,6 +137,10 @@ public class BulkEditNodesJob extends AbstractJob{
 			replaceToken = prepareParam(context, "replaceToken", true);
 		}
 
+		if(mode.equals(Mode.Remove)){
+			searchToken = prepareParam(context, "searchToken", true);
+		}
+
 		lucene =prepareParam(context, "lucene", false);
 
 		startFolder =prepareParam(context, "startFolder", true);
@@ -172,7 +176,28 @@ public class BulkEditNodesJob extends AbstractJob{
 				nodeService.setProperty(nodeRef,QName.createQName(property),value);
 			}
 			else if(mode.equals(Mode.Remove)){
-				nodeService.removeProperty(nodeRef,QName.createQName(property));
+				if(searchToken != null){
+					Serializable current = nodeService.getProperty(nodeRef,QName.createQName(property));
+					if(current != null){
+						boolean remove = false;
+						if(current instanceof String){
+							if(searchToken.equals((String)current)){
+								remove = true;
+							}
+						}else if(current instanceof List){
+							for(Object o : (List)current){
+								if(searchToken.equals(0)){
+									remove = true;
+								}
+							}
+						}
+						if(remove){
+							nodeService.removeProperty(nodeRef, QName.createQName(property));
+						}
+					}
+				}else {
+					nodeService.removeProperty(nodeRef, QName.createQName(property));
+				}
 			} else if(mode.equals(Mode.ReplaceToken) || mode.equals(Mode.ReplaceMapping)) {
 				Serializable current = nodeService.getProperty(nodeRef, QName.createQName(property));
 				if (current != null) {
